refactor(Icons): move stories inside component folder
