+++ src/components/Icons/stories/index.tsx
@@ -3,9 +3,9 @@ import { View } from 'react-native';
 import { storiesOf } from '@storybook/react-native';
 import { select, number } from '@storybook/addon-knobs';
 
-import * as icons from '@components/Icons';
+import * as icons from '..';
 import { colors } from '@magnetis/astro-galaxy-tokens';
-import { colorOptions, gradientOptions } from './options';
+import { colorOptions, gradientOptions } from '@root/storybook/options';
 
 const iconsStories = storiesOf('Icons', module).addDecorator((Story: any) => (
   <View style={{ alignSelf: 'center', alignItems: 'center', justifyContent: 'center', flex: 1 }}>
@@ -25,3 +25,5 @@ Object.keys(icons).forEach((key) => {
     />
   ));
 });
+
+export {};
