+++ src/stories/index.ts
@@ -1,5 +1,5 @@
 import './Buttons.stories';
 import './Text.stories';
 import './ControlsToggles.stories';
-import './Icons.stories';
+import '@components/Icons/stories';
 import '@components/Inputs/stories';
